+++ src/components/App.js
@@ -1,10 +1,15 @@
+import { Provider } from 'react-redux';
 import BooksList from '../containers/BooksList';
 import BooksForm from '../containers/BooksForm';
+import store from '../reducers/store';
 
 const App = () => (
   <div>
-    <BooksList />
-    <BooksForm />
+    <Provider store={store}>
+
+      <BooksList />
+      <BooksForm />
+    </Provider>
   </div>
 );
 export default App;
